I think were done here!

+++ src/main/java/hospital/service/HospitalService.java
@@ -85,7 +85,6 @@ public class HospitalService {
 		patient.setComplaint(hospitalPatient.getComplaint());
 		patient.setDiagnosis(hospitalPatient.getDiagnosis());
 	}
-
 	private Patient findOrCreatePatient(Long patientId, Long hospitalId) {
 			if (patientId == null){
 			System.out.println("no patient found!");
@@ -104,9 +103,18 @@ public class HospitalService {
 				return patient;
 			}
 		}throw new IllegalArgumentException("No such patient for hospital ID= " + hospitalId);
-		
 	}	
-
+	
+//	private Staff findStaffById(Long hospitalId, Long staffId) {
+//		Staff staff =  staffDao.findById(staffId).orElseThrow(() -> new NoSuchElementException("Hospital with ID= "
+//				+ hospitalId + " does not have an employee with employee ID= " + staffId + "."));
+//		
+//		if(staffId == staff.getHospital().getHospitalId()) {
+//			return staff;
+//		}else {
+//			throw new IllegalArgumentException("No such staff member ID for a hospitil with ID= " + hospitalId);
+//		}
+	
 	@Transactional(readOnly = false)
 	public HospitalStaff saveStaff(Long hospitalId, HospitalStaff hospitalStaff) {
 		Hospital hospital = findHospitalById(hospitalId);
@@ -182,7 +190,7 @@ public class HospitalService {
 
 	private Patient findPatientById (Long patientId) {
 		return patientDao.findById(patientId).orElseThrow(()
-				-> new NoSuchElementException("Hospital with ID = " 
+				-> new NoSuchElementException("Patient with ID = " 
 						+ patientId + " is not a valid ID number."));
 	}
 	
@@ -192,11 +200,26 @@ public class HospitalService {
 		Patient patient = optPatient.get();
 		return new HospitalPatient(patient);
 	}
+//	original
+//	@Transactional (readOnly = false)
+//	public void deletePatientById(Long patientId, Long hospitalId) {
+//		Patient patient = findPatientById(patientId);
+//		patientDao.delete(patient);
+//	}
 	
-	@Transactional (readOnly = false)
-	public void deletePatientById(Long patientId) {
+
+	
+		@Transactional (readOnly = false)
+	public void deletePatientById( Long hospitalId, Long patientId ) {
 		Patient patient = findPatientById(patientId);
+		Hospital hospital = findHospitalById(hospitalId);
+		
+		if(patientId == patient.getPatientId() && hospitalId == hospital.getHospitalId()) {
 		patientDao.delete(patient);
+		}
+		else {
+			throw new IllegalArgumentException("No such patient ID for a hospitil with ID= " + hospitalId);
+		}
 	}
 	
 	@Transactional (readOnly = true)
